+++ Reinvest/Verification/src/Port/Api/RegistrationApiType.ts
@@ -1,10 +1,13 @@
 import { ContainerInterface } from 'Container/Container';
 import { VerifyAccount } from 'Verification/IntegrationLogic/UseCase/VerifyAccount';
+import { AdminVerificationActions } from 'Verification/Port/Api/AdminVerificationActions';
 
 export type VerificationApiType = {
+  recoverVerification: AdminVerificationActions['recoverVerification'];
   verifyAccount: VerifyAccount['verify'];
 };
 
 export const verificationApi = (container: ContainerInterface): VerificationApiType => ({
+  recoverVerification: container.delegateTo(AdminVerificationActions, 'recoverVerification'),
   verifyAccount: container.delegateTo(VerifyAccount, 'verify'),
 });