+++ Reinvest/Verification/src/Providers/PortsProvider.ts
@@ -3,6 +3,7 @@ import { RegistrationService } from 'Verification/Adapter/Modules/RegistrationSe
 import { Verification } from 'Verification/index';
 import { VerifyAccount } from 'Verification/IntegrationLogic/UseCase/VerifyAccount';
 import { VerifierRepository } from 'Verification/IntegrationLogic/Verifier/VerifierRepository';
+import { AdminVerificationActions } from 'Verification/Port/Api/AdminVerificationActions';
 
 export class PortsProvider {
   private config: Verification.Config;
@@ -13,6 +14,7 @@ export class PortsProvider {
 
   public boot(container: ContainerInterface) {
     // api
+    container.addSingleton(AdminVerificationActions, [VerifierRepository]);
     container.addSingleton(VerifyAccount, [RegistrationService, VerifierRepository]);
   }
 }